feat: create pages folder to set up page routing
